First Come First Serve is finished
There are some un-noticeable bugs that I doubt will make a difference.
Everything to run the program I have written, so all I need is to write
the algorithms for Shortest Job First, and apply Round robin that romeo
is writing.  The current structure should allow those to be easily
added into classes using the FCFS as a template.